Deletes message class and redistrubtes messages to their corresponding methods
