+++ README.md
@@ -1,4 +1,5 @@
 # Battleship
-## A game by the *DREAM TEAM* Kate Tester and Carson Jardine
+## A game by the *DREAM TEAM*:
+### Kate Tester and Carson Jardine
 
 ![](https://gph.is/2OYabGQ)
